+++ NEWS.md
@@ -9,6 +9,7 @@ git2rdata 0.0.4 (2019-05-03)
 
   * `read_vc()` checks the meta data hash. A mismatch results in an error.
   * The meta data gains a data hash. A mismatch throws a warning. This allows the data to up updated by other software.
+  * `is_git2rmeta()` validates metadata
   
 ### Bugfixes